requirements.txt
torch
torchvision
transformers
opencv-python
numpy
PyYAML
timm 
